Adding new method for creating context, issue #63
